+++ grand_trade_auto/database/postgres.py
@@ -33,18 +33,17 @@ class Postgres(database_meta.Database):
       _host (str): The host URL.
       _post (int): The port number on that host for accessing the database.
       _database (str): The database to open.
+      _user (str): The user to use for authentication.
+      _password (str): The password to use for authentication.
 
       _conn (connection): The cached database connection; or None if not
         connected and cached.
 
       [inherited from Database]:
         _env (str): The run environment type valid for using this database.
-        _cp_db_id (str): The id used as the section name in the database conf.
-          Will be used for loading credentials on-demand.
-        _cp_secrets_id (str): The id used as the section name in the secrets
-          conf.  Will be used for loading credentials on-demand.
+        _db_id (str): The id used as the section name in the database conf.
     """
-    def __init__(self, host, port, database, **kwargs):
+    def __init__(self, host, port, database, user, password, **kwargs):
         """
         Creates the database handle.
 
@@ -52,6 +51,8 @@ class Postgres(database_meta.Database):
           host (str): The host URL.
           post (int): The port number on that host for accessing the database.
           database (str): The database to open.
+          user (str): The user to use for authentication.
+          password (str): The password to use for authentication.
 
           See parent(s) for required kwargs.
         """
@@ -59,13 +60,15 @@ class Postgres(database_meta.Database):
         self._host = host
         self._port = port
         self._database = database
+        self._user = user
+        self._password = password
 
         self._conn = None
 
 
 
     @classmethod
-    def load_from_config(cls, db_cp, db_id, secrets_id):
+    def load_from_config(cls, db_cp, db_id):
         """
         Loads the database config for this database from the configparsers
         from files provided.
@@ -74,21 +77,28 @@ class Postgres(database_meta.Database):
           db_cp (configparser): The full configparser from the database conf.
           db_id (str): The ID name for this database as it appears as the
             section header in the db_cp.
-          secrets_id (str): The ID name for this database's secrets as it
-            appears as the section header in the secrets_cp.
 
         Returns:
           db (Postgres): The Postgres object created and loaded from config
             based on the provided config data.
         """
-        kwargs = {}
+        db_cp = config.read_conf_file('databases.conf')
+        secrets_cp = config.read_conf_file('.secrets.conf')
+        secrets_id = config.get_matching_secrets_id(secrets_cp, 'database',
+                db_id)
 
+        kwargs = {}
         kwargs['env'] = db_cp[db_id]['env']
-        kwargs['cp_db_id'] = db_id
-        kwargs['cp_secrets_id'] = secrets_id
+        kwargs['db_id'] = db_id
         kwargs['host'] = db_cp[db_id]['host url']
         kwargs['database'] = db_cp[db_id]['database']
         kwargs['port'] = db_cp.getint(db_id, 'port', fallback=5432)
+
+        kwargs['user'] = db_cp.get(db_id, 'username', fallback=None)
+        kwargs['user'] = secrets_cp.get(secrets_id, 'username',
+                fallback=kwargs['user'])
+        kwargs['password'] = secrets_cp.get(secrets_id, 'password',
+                fallback=None)
 
         db = Postgres(**kwargs)
         return db
@@ -127,23 +137,15 @@ class Postgres(database_meta.Database):
         if cache and self._conn is not None:
             return self._conn
 
-        db_cp = config.read_conf_file('databases.conf')
-        secrets_cp = config.read_conf_file('.secrets.conf')
-
         kwargs = {
             'host': self._host,
             'port': self._port,
+            'user': self._user,
+            'password': self._password,
         }
         if database is None:
             database = self._database
         kwargs['database'] = database
-
-        kwargs['user'] = db_cp.get(self._cp_db_id, 'username',
-                fallback=None)
-        kwargs['user'] = secrets_cp.get(self._cp_secrets_id, 'username',
-                fallback=kwargs['user'])
-        kwargs['password'] = secrets_cp.get(self._cp_secrets_id, 'password',
-                fallback=None)
 
         conn = psycopg2.connect(**kwargs)
         if cache: